Fix amount serialized as string instead of number
Pydantic serializes Decimal to a JSON string, causing NaN in the
frontend when arithmetic is performed. Use float instead so amount
is always a JSON number.

+++ backend/models.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from datetime import date, datetime, time
-from decimal import Decimal
 from typing import Literal
 
 from pydantic import BaseModel
@@ -23,7 +22,7 @@ class ScanResult(BaseModel):
     expense_id: int
     image_id: int
     image_url: str
-    amount: Decimal | None
+    amount: float | None
     currency: str
     date: date | None
     time: time | None
@@ -40,7 +39,7 @@ class ExpenseResponse(BaseModel):
     id: int
     image_id: int
     image_url: str
-    amount: Decimal | None
+    amount: float | None
     currency: str
     date: date | None
     time: time | None
